organise files properly
