+++ frontend/src/app/app.component.ts
@@ -4,7 +4,7 @@ import {MatToolbarModule} from "@angular/material/toolbar";
 import {MatButtonModule} from "@angular/material/button";
 import {MatIconModule} from "@angular/material/icon";
 import {MatSidenavModule} from "@angular/material/sidenav";
-import {NavbarComponent} from "./navbar/navbar.component";
+import {NavbarComponent} from "./shared/layout/navbar/navbar.component";
 
 @Component({
   standalone: true,
