+++ frontend/src/app/features/gigs/event-card/event-card.component.html
