+++ frontend/src/app/shared/layout/navbar/navbar.component.ts
