+++ frontend/src/app/shared/models/gig.ts
